fix(recv_cat): use `ssize_t` for all buffer sizes
Removes narrowing conversion which causes an implementation-defined warning.

+++ src/utils/hyprland_socket_handling.c
@@ -112,7 +112,7 @@ int set_up_hyprland_socket(Socket socket_type, SocketData *socket_data) {
 
 // Looped version of recv that dynamically allocates a growing chunk of memory to store data
 // retrieved.
-char *recv_cat(int file_descriptor, size_t buffer_size, int flags) {
+char *recv_cat(int file_descriptor, ssize_t buffer_size, int flags) {
     char *full_data = (char *)calloc(1, buffer_size);
     if (full_data == NULL) {
         perror("calloc");
@@ -120,7 +120,7 @@ char *recv_cat(int file_descriptor, size_t buffer_size, int flags) {
     }
     ssize_t signed_buffer_size = buffer_size;
     ssize_t num_bytes_received = buffer_size;
-    int cur_buffer_size = 0;
+    ssize_t cur_buffer_size = 0;
 
     while (num_bytes_received == signed_buffer_size) {
         char buffer[buffer_size];
